+++ modules/project/import-ui/src/sap/commerce/toolset/project/wizard/InformationStep.kt
@@ -92,7 +92,7 @@ class InformationStep(wizardContext: WizardContext) : ProjectImportWizardStep(wi
         separator()
 
         row {
-            label("JRebel && HotSwap")
+            label("JRebel & HotSwap")
                 .bold()
                 .align(AlignX.CENTER)
         }