+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ## [2025.2.4.4]
 
 #### Release contributors
-- 2 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.2.4.4+author%3Amlytvyn+is%3Apr)
+- 3 PR(s) by [Mykhailo Lytvyn](https://github.com/epam/sap-commerce-intellij-idea-plugin/pulls?q=milestone%3A2025.2.4.4+author%3Amlytvyn+is%3Apr)
 
 ### `Project Import` enhancements
 - Updated project import "Information" step [#1612](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1612)
+- Updated project import "Project Details" step [#1614](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1614)
 
 ### `CCv2` enhancements
 - Renew `fetch` icon right after completing the request [#1611](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1611)